moved the test menu to a screen
