--- src/testMode/index.ts
@@ -1,2 +1,0 @@
-export * from './TestMode';
-export * from './DemoMode';
